Fix test failure with bytestring 0.9
The Show instance for lazy ByteStrings in 0.9 includes Chunk and Empty
garbage, which caused the tests to fail. This hacky solution is to
convert the lazy ByteString into a strict ByteString. There might be
better solutions, but I couldn't think of any.

Note that conditional compilation doesn't seem to play well with
doctests.