Properly support bytestring-0.11 (#53)
It was pointed out in ekmett/bytes#52
that my patch to allow `bytes` to build with `bytestring-0.11.*` in
ed6135e1b59da7b46407d3fdc9b161810bac8baa was incomplete. This is because I
overlooked the fact that `bytestring-0.11.*` only exports the `PS` pattern
synonym on GHC 8.0 or later, and since `bytes` matches on `PS`, it wouldn't
compile if you combined `bytestring-0.11` with a pre-8.0 version of GHC.

I adopted a similar fix as in
https://github.com/ekmett/lens/commit/1debced4e9cb8b904b1f52c3ab10a551e40f3b2d#diff-700f6fcfab7d955dd35e76a01956c66b89cff9f681274163b6d60051229ef14dR217
to allow `bytestring-0.11` to be used with all supported versions of GHC.

Fixes #52.